+++ middleware-workflow-module/workflow-model/src/main/java/co/tpg/workflow/function/response/WorkflowListResponse.java
@@ -1,0 +1,12 @@
+package co.tpg.workflow.function.response;
+
+import co.tpg.workflow.function.model.Workflow;
+import java.util.List;
+
+/**
+ * Response class for List of workflows
+ * @author Andrej
+ * @since 2019-10-12
+ */
+public class WorkflowListResponse extends AbstractResponse<List<Workflow>> {
+}
